src/RAG/requirements.txt
# === Core LangChain ecosystem ===
langchain==1.0.3
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0

# === Vector store / retriever ===
chromadb>=0.5.5
faiss-cpu>=1.8.0.post1        # (optional, 다른 벡터 DB와 병행 가능)

# === Embeddings (HuggingFace + KURE-v1) ===
sentence-transformers>=3.0.1
torch>=2.3.0
transformers>=4.43.0

# === Web / API server ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# === Data handling / preprocessing ===
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
scikit-learn>=1.3.0

# === Utility / reranking / fuzzy matching ===
rapidfuzz>=3.9.0
tqdm>=4.66.4

# === Optional (for local dev) ===
requests>=2.31.0
pydantic>=2.7.4
